Extract multi swap quotes into separate classes
