+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/Providers/OneInchMultiSwapProvider.swift
@@ -87,7 +87,7 @@ class OneInchMultiSwapProvider: BaseEvmMultiSwapProvider {
             )
         }
 
-        return ConfirmationQuote(
+        return OneInchMultiSwapConfirmationQuote(
             quote: quote,
             swap: resolvedSwap,
             insufficientFeeBalance: insufficientFeeBalance,
@@ -105,7 +105,7 @@ class OneInchMultiSwapProvider: BaseEvmMultiSwapProvider {
     }
 
     override func swap(tokenIn: MarketKit.Token, tokenOut _: MarketKit.Token, amountIn _: Decimal, quote: IMultiSwapConfirmationQuote) async throws {
-        guard let quote = quote as? ConfirmationQuote else {
+        guard let quote = quote as? OneInchMultiSwapConfirmationQuote else {
             throw SwapError.invalidQuote
         }
 
@@ -133,7 +133,7 @@ class OneInchMultiSwapProvider: BaseEvmMultiSwapProvider {
         try OneInchKit.Kit.routerAddress(chain: chain)
     }
 
-    private func internalQuote(tokenIn: MarketKit.Token, tokenOut: MarketKit.Token, amountIn: Decimal) async throws -> Quote {
+    private func internalQuote(tokenIn: MarketKit.Token, tokenOut: MarketKit.Token, amountIn: Decimal) async throws -> OneInchMultiSwapQuote {
         let blockchainType = tokenIn.blockchainType
         let chain = evmBlockchainManager.chain(blockchainType: blockchainType)
 
@@ -152,7 +152,7 @@ class OneInchMultiSwapProvider: BaseEvmMultiSwapProvider {
             amount: amount
         )
 
-        return await Quote(
+        return await OneInchMultiSwapQuote(
             quote: quote,
             recipient: storage.recipient(blockchainType: blockchainType),
             slippage: slippage,
@@ -186,147 +186,6 @@ extension OneInchMultiSwapProvider {
         case invalidSwap
         case noEvmKitWrapper
         case noGasLimit
-    }
-}
-
-extension OneInchMultiSwapProvider {
-    class Quote: BaseEvmMultiSwapProvider.Quote {
-        let quote: OneInchKit.Quote
-        let recipient: Address?
-        let slippage: Decimal
-
-        init(quote: OneInchKit.Quote, recipient: Address?, slippage: Decimal, allowanceState: MultiSwapAllowanceHelper.AllowanceState) {
-            self.quote = quote
-            self.recipient = recipient
-            self.slippage = slippage
-
-            super.init(allowanceState: allowanceState)
-        }
-
-        override var amountOut: Decimal {
-            quote.amountOut ?? 0
-        }
-
-        override var settingsModified: Bool {
-            super.settingsModified || recipient != nil || slippage != MultiSwapSlippage.default
-        }
-
-        override func fields(tokenIn: MarketKit.Token, tokenOut: MarketKit.Token, currency: Currency, tokenInRate: Decimal?, tokenOutRate: Decimal?) -> [MultiSwapMainField] {
-            var fields = super.fields(tokenIn: tokenIn, tokenOut: tokenOut, currency: currency, tokenInRate: tokenInRate, tokenOutRate: tokenOutRate)
-
-            if let recipient {
-                fields.append(
-                    MultiSwapMainField(
-                        title: "swap.recipient".localized,
-                        value: recipient.title,
-                        valueLevel: .regular
-                    )
-                )
-            }
-
-            if slippage != MultiSwapSlippage.default {
-                fields.append(
-                    MultiSwapMainField(
-                        title: "swap.slippage".localized,
-                        value: "\(slippage.description)%",
-                        valueLevel: MultiSwapSlippage.validate(slippage: slippage).valueLevel
-                    )
-                )
-            }
-
-            return fields
-        }
-
-        override func cautions() -> [CautionNew] {
-            var cautions = super.cautions()
-
-            switch MultiSwapSlippage.validate(slippage: slippage) {
-            case .none: ()
-            case let .caution(caution): cautions.append(caution.cautionNew(title: "swap.advanced_settings.slippage".localized))
-            }
-
-            return cautions
-        }
-    }
-
-    class ConfirmationQuote: BaseEvmMultiSwapProvider.ConfirmationQuote {
-        let quote: Quote
-        let swap: Swap?
-        let insufficientFeeBalance: Bool
-
-        init(quote: Quote, swap: Swap?, insufficientFeeBalance: Bool, evmFeeData: EvmFeeData?, nonce: Int?) {
-            self.quote = quote
-            self.swap = swap
-            self.insufficientFeeBalance = insufficientFeeBalance
-
-            super.init(gasPrice: swap?.transaction.gasPrice, evmFeeData: evmFeeData, nonce: nonce)
-        }
-
-        override var amountOut: Decimal {
-            swap?.amountOut ?? quote.quote.amountOut ?? 0
-        }
-
-        override var canSwap: Bool {
-            super.canSwap && swap != nil && !insufficientFeeBalance
-        }
-
-        override func cautions(feeToken: MarketKit.Token?) -> [CautionNew] {
-            var cautions = super.cautions(feeToken: feeToken)
-
-            if insufficientFeeBalance {
-                cautions.append(
-                    .init(
-                        title: "fee_settings.errors.insufficient_balance".localized,
-                        text: "ethereum_transaction.error.insufficient_balance_with_fee".localized(feeToken?.coin.code ?? ""),
-                        type: .error
-                    )
-                )
-            }
-
-            cautions.append(contentsOf: quote.cautions())
-
-            return cautions
-        }
-
-        override func priceSectionFields(tokenIn: MarketKit.Token, tokenOut: MarketKit.Token, feeToken: MarketKit.Token?, currency: Currency, tokenInRate: Decimal?, tokenOutRate: Decimal?, feeTokenRate: Decimal?) -> [SendConfirmField] {
-            var fields = super.priceSectionFields(tokenIn: tokenIn, tokenOut: tokenOut, feeToken: feeToken, currency: currency, tokenInRate: tokenInRate, tokenOutRate: tokenOutRate, feeTokenRate: feeTokenRate)
-
-            if let recipient = quote.recipient {
-                fields.append(
-                    .address(
-                        title: "swap.recipient".localized,
-                        value: recipient.title,
-                        blockchainType: tokenOut.blockchainType
-                    )
-                )
-            }
-
-            let slippage = quote.slippage
-
-            if slippage != MultiSwapSlippage.default {
-                fields.append(
-                    .levelValue(
-                        title: "swap.slippage".localized,
-                        value: "\(slippage.description)%",
-                        level: MultiSwapSlippage.validate(slippage: slippage).valueLevel
-                    )
-                )
-            }
-
-            let minAmountOut = amountOut * (1 - slippage / 100)
-
-            fields.append(
-                .value(
-                    title: "swap.confirmation.minimum_received".localized,
-                    description: nil,
-                    coinValue: CoinValue(kind: .token(token: tokenOut), value: minAmountOut),
-                    currencyValue: tokenOutRate.map { CurrencyValue(currency: currency, value: minAmountOut * $0) },
-                    formatFull: true
-                )
-            )
-
-            return fields
-        }
     }
 }
